changes made

+++ Pro_mini.py
@@ -41,7 +41,7 @@ class Pro_mini():
         try:
             response = self.serial.readline().decode('utf-8').rstrip()
             if len(response) != 0:
-                return int(response)
+                return float(response)
             else:
                 return self.read_Response()
         except serial.serialutil.SerialException:
